refactor agents from fencing, using cluster-fence-config and modified comments and readme.
- using cluster-fence-config
- adding more comments
- change readme to fit current behavior
- creating utils.go and agents.go to simplify the code structure
