+++ demo/demo-kubernetes-minion-group.yaml
@@ -11,6 +11,7 @@ items:
    config.properties: |-
     grace_timeout=10
     giveup_retries=5
+    roles=
 
 - kind: ConfigMap
   apiVersion: v1
